Develop

+++ express-webshop/routes/baskets.js
@@ -26,27 +26,23 @@ router.get('/', async (req, res) => {
   });
   actualQuantity = actualQuantity[0].totalQuantity;
 
+
   if (req.user.id) {
     res.render('baskets', {
       basketItemsWithNamesAndPrices: data,
       totalPrice: price,
       user: req.user || {},
       showQuantity: actualQuantity,
-
     });
-
   }
 });
 
 // a bejelentkezett user kosarának ürítése
 router.get('/empty/:userid', async (req, res) => {
   database.deleteRecord('baskets', {
-    userid: req.user.id
+    userid: req.user.id,
   });
-  res.render('baskets', {
-    user: req.user || {},
-
-  });
+  res.redirect('/baskets');
 });
 
 router.post('/donate', async (req, res) => {
@@ -59,11 +55,13 @@ router.post('/donate', async (req, res) => {
   basket.forEach((el) => {
     quantitySum += el.quantity;
   });
-  await database.createRecord('orders', {
-    userid: req.user.id,
-    quantity: quantitySum,
-    status: 1
-  });
+  if (quantitySum > 0) {
+    await database.createRecord('orders', {
+      userid: req.user.id,
+      quantity: quantitySum,
+      status: 1,
+    });
+  }
   let orderID = await database.readRecord('orders', {
     userid: req.user.id,
     limit: 1,
@@ -87,21 +85,56 @@ router.post('/donate', async (req, res) => {
 
   basketItem.forEach((el) => {
     const balanceOfProjects = el.balance + el.donation * el.quantity;
+    const date = new Date();
     database.updateRecord('projects', {
       id: el.projectid,
     }, {
-      balance: balanceOfProjects
+      balance: balanceOfProjects,
+      lastfunded: `${date.getFullYear()}-${date.getMonth() + 1}-${date.getDate()}`,
     });
   });
-  await database.deleteRecord('baskets', {
-    userid: req.user.id,
+
+  const userAward = await database.readRecord('baskets', {
+    'users.id': req.user.id,
+    from: 'INNER JOIN users ON baskets.userid = users.id',
+    select: 'users.donations as donations, baskets.quantity as quantity, users.points',
   });
-  res.redirect('/thankyou');
+
+  const price = await database.readRecord('projects', {
+    userid: req.user.id,
+    select: 'SUM(projects.donation*baskets.quantity) as amount',
+    from: 'INNER JOIN baskets ON projects.id = baskets.projectid',
+  });
+
+  const donationsPerUser = price[0].amount;
+
+  userAward.forEach((el) => {
+    const amountOfDonations = el.donations + donationsPerUser;
+    const amountOfUsersPoints = amountOfDonations / 5
+
+    if (amountOfDonations > 0) {
+      database.updateRecord('users', {
+        id: req.user.id,
+      }, {
+        donations: amountOfDonations,
+        points: amountOfUsersPoints,
+      });
+    }
+  });
+
+
+  if (quantitySum > 0) {
+    await database.deleteRecord('baskets', {
+      userid: req.user.id,
+    });
+    res.redirect('/thankyou');
+  } else {
+    res.redirect('/baskets');
+  }
 });
 
 // post a project details oldalról
 router.post('/:id', async (req, res) => {
-  console.log('HEEEEEEEEEEEEE', req.params.id);
   const quantity = await database.readRecord('baskets', {
     userid: req.user.id || 0,
     projectid: req.params.id,
@@ -115,7 +148,6 @@ router.post('/:id', async (req, res) => {
       userid: req.user.id || 0,
       quantity: req.body.projectQuantity || 1,
     });
-
   } else {
     const incrementedQuantity = quantity[0].quantity + parseInt(req.body.projectQuantity, 10);
     await database.updateRecord('baskets', {
@@ -124,7 +156,6 @@ router.post('/:id', async (req, res) => {
     }, {
       quantity: incrementedQuantity,
     });
-
   }
   res.redirect('/baskets');
 });
@@ -160,6 +191,5 @@ router.post('/updateAdd/:id', async (req, res) => {
   });
   res.redirect('/baskets');
 });
-
 
 module.exports = router;
